pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentward"
version = "0.1.0"
description = "Open-source permission control plane for AI agents. Scan, enforce, and audit every tool call."
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "AgentWard Contributors" },
]
keywords = ["ai", "agents", "mcp", "security", "permissions", "governance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "aiohttp>=3.9",
]

[project.scripts]
agentward = "agentward.cli:app"

[project.urls]
Homepage = "https://agentward.ai"
Repository = "https://github.com/agentward-ai/agentward"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "A", "SIM"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
]
